refactor: Simplify step navigation and form submission logic in General and StepNavigation components
