+++ components/hackathons/project-submission/hooks/useSubmissionForm.ts
@@ -5,7 +5,6 @@ import { z } from 'zod';
 import axios from 'axios';
 import { useSession } from 'next-auth/react';
 import { useRouter } from 'next/navigation';
-import { toast } from '@/components/ui/use-toast';
 
 export const FormSchema = z.object({
   project_name: z
